isMutant function implementation It is requested to create the isMutant function that indicates if a dna matrix belongs to a mutant human or not.
What dna belongs to a mutant?
All human dna has an NxN matrix structure such that each element of the matrix has one of the four nitrogenous bases G, T, A or C as its value. A human is a mutant if his dna has more than one sequence followed by 4 nitrogenous bases from the same value in horizontal, vertical or oblique position within the respective matrix.

Technical details
The solution basically consisted of creating a service called `MutantIdentifier` that receives the dna as a string array and indicates with a boolean value if said `dna` received as a parameter belongs to a mutant human or not.
To identify mutant sequences, the `dna` data is first validated. First, it is verified that the dna is an array to later validate its content, which means validating that it is not empty, that its size is `NxN` and as each row of the array is valid, it is verified that it contains the valid nitrogenous base characters. Once the dna has been validated, the second consists of executing the algorithm.

The sequence identification algorithm basically has the three cases executed in order of complexity (horizontal, vertical and oblique), all supported by the identification of the number of sequences of size `4` of the same character within a string through the ` function performSequence` inside the `MutantLineSequenceCounter` class.

Horizontal
To identify the horizontal sequences, a simple traversal of the array is made such that for each element of the array, that is, each row of the matrix, the `performSequence` function mentioned above is executed.

Vertical
To identify the vertical sequences, the trasposed function of the StrMatrixHelper helper is used, which returns the transposed matrix of the matrix sent as a parameter, in this case the dna. With the transposed dna matrix, the vertical sequences are now in a horizontal position, allowing the algorithm mentioned in the previous paragraph to be executed.

Oblique
Finally, to identify the oblique sequences, the oblique sequences are read from the upper left part of the matrix to the lower right part such that the number of characters exceeds the minimum value 4. As this data is read, it is validated in concatenations of 4 characters the presence of sequences of the same value until finding 2 or more. This process is repeated with the same transposed matrix generated in the previous step.

Returning to the `MutantLineSequenceCounter` class, the protagonist of all the algorithms, it basically counts the number of matches between the string of each row in each generated array of the `3` cases, and an automatically generated regex from the allowed nitrogenous bases.