REST implementation (POST /mutant) It is requested to create a REST API to receive POST requests with the dna as body, with the response indicating whether or not the received dna belongs to a mutant. If it belongs to a mutant, the request status should be 200, otherwise it should be 403.
To build the API, the express package was used with its code generator. Later we created the `routers/mutant` that sends the post request to the create function of `controllers/mutant`. The controller was programmed in such a way that it calls the MutantIdentifier service so that, based on the response, the status of the response is modified or not and finally it is answered.