Database, GET /stats and unit tests
It is requested to add a database to register uniquely the dna requested to the REST service, in addition to responding to the GET /stats request to obtain the number of registered requests from humans as shown below, and unit tests to validate the correct application operation.

1. Database
MongoDB is chosen as the database engine as a result of an investigation whose conclusions are summarized [here](https://www.plesk.com/blog/various/mongodb-vs-postgresql/). The mongoose package configured through the `config/database.js` and `config/mongodb` files is integrated in support of the `dotenv` package for assigning credentials for connecting to the database.

Since this database engine requires no migrations, only the corresponding mongoose DnaSchema schema is created in `models/dna.js` in which the `sequence` and `isMutant` database attributes are assigned. The database sequence attribute records the dna in join with the `|` character. To automate and improve the interaction of the use of `sequence` and `dna` where the former is the latter produced as a string, the virtual attribute `dna` was created within the schema, facilitating and centralizing the use of these.

With the need for registration in DB, the `DnaClassifier` service is added, which slightly modifies the logic officially defined in the API. Now the controller consumes this service which initially instantiates the `Dna` model, proceeds to validate the dna. If the dna is invalid, directly set the `isMutant` attribute to false and respond. On the contrary, if the dna is valid, it is searched in the database and it is assigned if it has been found; if it was not found, the `MutantIdentifier` service is executed and later it makes the registration to the database.

2. stats controller
Added files and configuration for `routes/stats` and `controllers/stats`. In this second file, it is executed on the `index` function, the `StatsService` service that queries the database through the `Dna` class and the `count()` function, memorizing the results for the `ratio` attribute. which is part of the answer. This service supports the `helpers/numbers_helper.js` helper to round the result of the `ratio` attribute.

3. Tests
The `supertest` and `jest` packages were integrated since they are the most used and their structure is easy to understand. Tests were made on the REST service (controllers) with all the cases that trigger the execution of multiple scenarios within services, helpers, models, etc.

Additionally, the response for case 404 was added.